Changed font size to float so smaller units can be used.

+++ commonSrc/net/xuset/tSquare/imaging/IFont.java
@@ -3,5 +3,5 @@ package net.xuset.tSquare.imaging;
 public interface IFont {
 	String getName();
 	TsTypeFace getTypeFace();
-	int getSize();
+	float getSize();
 }
