+++ commonSrc/net/xuset/tSquare/imaging/TsFont.java
@@ -2,10 +2,10 @@ package net.xuset.tSquare.imaging;
 
 public class TsFont implements IFont{
 	private final String name;
-	private final int size;
+	private final float size;
 	private final TsTypeFace typeFace;
 	
-	public TsFont(String name, int size, TsTypeFace typeFace) {
+	public TsFont(String name, float size, TsTypeFace typeFace) {
 		this.name = name;
 		this.size = size;
 		this.typeFace = typeFace;
@@ -26,7 +26,7 @@ public class TsFont implements IFont{
 	}
 
 	@Override
-	public int getSize() {
+	public float getSize() {
 		return size;
 	}
 }
